Align scatter_reduce benchmark suite with operator ids
