+++ benchmark/test_scatter_reduce.py
@@ -26,22 +26,6 @@ class TensorSelectBenchmark(base.GenericBenchmark2DOnly):
         return shapes
 
 
-class ScatterReducePublicBenchmark(base.GenericBenchmark2DOnly):
-    def set_more_metrics(self):
-        return ["gbps"]
-
-    def set_more_shapes(self):
-        if flag_gems.vendor_name == "kunlunxin":
-            return []
-
-        shapes = super().set_more_shapes()
-        return [
-            shape
-            for shape in shapes
-            if len(shape) == 2 and shape[0] > 16 and shape[1] > 16
-        ]
-
-
 def gather_input_fn(shape, dtype, device):
     inp = torch.randn(shape, dtype=dtype, device=device)
 
@@ -84,22 +68,6 @@ def scatter_inplace_input_fn_factory(reduce=None):
     return inner
 
 
-def scatter_reduce_public_input_fn_factory(reduce, include_self=True, out=False):
-    def inner(shape, dtype, device):
-        inp = torch.randn(shape, dtype=dtype, device=device)
-        dim = -1
-        src_shape = list(shape)
-        src_shape[dim] = max(1, src_shape[dim] // 2)
-        src = torch.randn(src_shape, dtype=dtype, device=device)
-        index = torch.randint(0, shape[dim], src_shape, dtype=torch.long, device=device)
-        kwargs = {"reduce": reduce, "include_self": include_self}
-        if out:
-            kwargs["out"] = torch.empty_like(inp)
-        yield inp, dim, index, src, kwargs
-
-    return inner
-
-
 def gather_scatter_gbps(bench_fn_args, latency):
     inp, dim, index = bench_fn_args[:3]
     data_shape = list(inp.shape)
@@ -127,55 +95,6 @@ def test_scatter_reduce_prod():
         op_name="scatter_reduce",
         torch_op=torch.scatter_reduce,
         input_fn=scatter_input_fn_factory("prod"),
-        get_gbps=gather_scatter_gbps,
-        dtypes=[torch.float16, torch.float32],
-    )
-    bench.run()
-
-
-@pytest.mark.scatter_reduce
-@pytest.mark.parametrize(
-    "reduce, include_self",
-    [("sum", True), ("mean", False), ("prod", True), ("amax", False), ("amin", True)],
-)
-def test_scatter_reduce_public_api(reduce, include_self):
-    bench = ScatterReducePublicBenchmark(
-        op_name="scatter_reduce_two",
-        torch_op=torch.scatter_reduce,
-        input_fn=scatter_reduce_public_input_fn_factory(reduce, include_self),
-        get_gbps=gather_scatter_gbps,
-        dtypes=[torch.float16, torch.float32],
-    )
-    bench.run()
-
-
-@pytest.mark.scatter_reduce_
-@pytest.mark.parametrize(
-    "reduce, include_self",
-    [("sum", True), ("mean", True), ("prod", True), ("amax", True), ("amin", True)],
-)
-def test_scatter_reduce_public_api_inplace(reduce, include_self):
-    bench = ScatterReducePublicBenchmark(
-        op_name="scatter_reduce_two_",
-        torch_op=torch.Tensor.scatter_reduce_,
-        input_fn=scatter_reduce_public_input_fn_factory(reduce, include_self),
-        get_gbps=gather_scatter_gbps,
-        dtypes=[torch.float16, torch.float32],
-        is_inplace=True,
-    )
-    bench.run()
-
-
-@pytest.mark.scatter_reduce
-@pytest.mark.parametrize(
-    "reduce, include_self",
-    [("sum", False), ("mean", True), ("amax", True), ("amin", False)],
-)
-def test_scatter_reduce_public_api_out(reduce, include_self):
-    bench = ScatterReducePublicBenchmark(
-        op_name="scatter_reduce_two_out",
-        torch_op=torch.scatter_reduce,
-        input_fn=scatter_reduce_public_input_fn_factory(reduce, include_self, out=True),
         get_gbps=gather_scatter_gbps,
         dtypes=[torch.float16, torch.float32],
     )
